infer types in rules on write
