+++ grakn-core/src/main/java/ai/grakn/graql/admin/Conjunction.java
@@ -18,7 +18,6 @@
 
 package ai.grakn.graql.admin;
 
-import ai.grakn.GraknTx;
 import javax.annotation.CheckReturnValue;
 import java.util.Set;
 
@@ -35,10 +34,4 @@ public interface Conjunction<T extends PatternAdmin> extends PatternAdmin {
      */
     @CheckReturnValue
     Set<T> getPatterns();
-
-    /**
-     * @return the corresponding reasoner query
-     */
-    @CheckReturnValue
-    ReasonerQuery toReasonerQuery(GraknTx tx);
 }
