+++ grakn-graql/src/main/java/ai/grakn/graql/internal/pattern/ConjunctionImpl.java
@@ -18,18 +18,13 @@
 
 package ai.grakn.graql.internal.pattern;
 
-import ai.grakn.GraknTx;
 import ai.grakn.graql.Var;
 import ai.grakn.graql.admin.Conjunction;
 import ai.grakn.graql.admin.Disjunction;
 import ai.grakn.graql.admin.PatternAdmin;
-import ai.grakn.graql.admin.ReasonerQuery;
 import ai.grakn.graql.admin.VarPatternAdmin;
-import ai.grakn.graql.internal.reasoner.query.ReasonerQueries;
-import ai.grakn.kb.internal.EmbeddedGraknTx;
 import com.google.auto.value.AutoValue;
 import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
 
 import java.util.List;
@@ -78,13 +73,6 @@ abstract class ConjunctionImpl<T extends PatternAdmin> extends AbstractPattern i
     @Override
     public Conjunction<?> asConjunction() {
         return this;
-    }
-
-    @Override
-    public ReasonerQuery toReasonerQuery(GraknTx tx){
-        Conjunction<VarPatternAdmin> pattern = Iterables.getOnlyElement(getDisjunctiveNormalForm().getPatterns());
-        // TODO: This cast is unsafe - this method should accept an `EmbeddedGraknTx`
-        return ReasonerQueries.create(pattern, (EmbeddedGraknTx<?>) tx);
     }
 
     private static <U extends PatternAdmin> Conjunction<U> fromConjunctions(List<Conjunction<U>> conjunctions) {
