+++ grakn-graql/src/main/java/ai/grakn/graql/internal/reasoner/rule/RuleUtils.java
@@ -23,7 +23,6 @@ import ai.grakn.concept.Rule;
 import ai.grakn.concept.SchemaConcept;
 import ai.grakn.graql.internal.reasoner.atom.Atom;
 import ai.grakn.graql.internal.reasoner.query.ReasonerQueryImpl;
-import ai.grakn.kb.internal.EmbeddedGraknTx;
 import ai.grakn.util.Schema;
 import com.google.common.base.Equivalence;
 
@@ -71,15 +70,7 @@ public class RuleUtils {
     public static Stream<Rule> getRulesWithType(SchemaConcept type, boolean direct, GraknTx graph){
         if (type == null) return getRules(graph);
         Set<SchemaConcept> parentTypes = direct? Sets.newHashSet(type) : type.subs().collect(Collectors.toSet());
-        Stream<Rule> typedRules = parentTypes.stream().flatMap(SchemaConcept::getRulesOfConclusion);
-
-        Stream<Rule> untypedRules = getRules(graph)
-                .filter(r -> !r.getConclusionTypes().findFirst().isPresent())
-                .filter(r -> {
-                    SchemaConcept headType = new InferenceRule(r, (EmbeddedGraknTx<?>) graph).getHead().getAtom().getSchemaConcept();
-                    return parentTypes.stream().anyMatch(t -> t.equals(headType));
-                });
-        return Stream.concat(typedRules, untypedRules);
+        return parentTypes.stream().flatMap(SchemaConcept::getRulesOfConclusion);
     }
 
     /**
